StudentOverflow/requirements.txt
flask>=3.0
supabase>=2.8.1
python-dotenv>=1.0.1
werkzeug>=3.0